--- backend/src/main/java/com/CS157Aproject/backend/BackendApplication.java
@@ -1,13 +1,0 @@
-package com.CS157Aproject.backend;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BackendApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BackendApplication.class, args);
-	}
-
-}
