--- backend/src/test/java/com/CS157Asection09team5/backend/BackendApplicationTests.java
@@ -1,12 +1,0 @@
-package com.CS157Asection09team5.backend;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BackendApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-}
