+++ server/src/main/java/com/CS157Ateam5/server/ServerApplication.java
@@ -1,0 +1,13 @@
+package com.CS157Ateam5.server;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ServerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ServerApplication.class, args);
+	}
+
+}
