+++ server/src/main/resources/application.properties
@@ -1,0 +1,4 @@
+spring.datasource.url=jdbc:mysql://localhost:3306/cs157a?useSSL=false&serverTimezone=UTC
+spring.datasource.username=root
+spring.datasource.password=munkie
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
