+++ server/src/test/java/com/CS157Ateam5/server/ServerApplicationTests.java
@@ -1,0 +1,13 @@
+package com.CS157Ateam5.server;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ServerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
